# api/openapi.yaml
openapi: 3.0.3
info:
  title: Incident Management API
  description: |
    REST API for managing IT incidents, owners, and optional GitHub repository integration.
    Provides full CRUD operations along with filtering, searching, and statistics endpoints.
  version: 1.0.0
  contact:
    name: API Support
    email: support@incident-now.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.incident-now.io/v1
    description: Production server
...
paths:
...
  /incidents:
    get:
      tags:
        - Incidents
      summary: List all incidents
      description: Retrieve a paginated list of incidents with optional filtering
      operationId: listIncidents
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - name: status
          in: query
          description: Filter by incident status
          schema:
            $ref: '#/components/schemas/IncidentStatus'
        - name: priority
          in: query
          description: Filter by priority level
          schema:
            $ref: '#/components/schemas/Priority'
        - name: severity
          in: query
          description: Filter by severity level
          schema:
            $ref: '#/components/schemas/Severity'
        - name: category
          in: query
          description: Filter by incident category
          schema:
            $ref: '#/components/schemas/IncidentCategory'
        - name: ownerId
          in: query
          description: Filter by owner ID
          schema:
            type: string
            format: uuid
        - name: assigneeId
          in: query
          description: Filter by assignee ID
          schema:
            type: string
            format: uuid
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
          example: "network,critical"
        - name: createdAfter
          in: query
          description: Filter incidents created after this date
          schema:
            type: string
            format: date-time
        - name: createdBefore
          in: query
          description: Filter incidents created before this date
          schema:
            type: string
            format: date-time
        - name: resolvedAfter
          in: query
          description: Filter incidents resolved after this date
          schema:
            type: string
            format: date-time
        - name: resolvedBefore
          in: query
          description: Filter incidents resolved before this date
          schema:
            type: string
            format: date-time
        - name: hasGithubRepo
          in: query
          description: Filter incidents with/without GitHub repository link
          schema:
            type: boolean
        - name: search
          in: query
          description: Search in title, description, and affected systems
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Incidents
      summary: Create a new incident
      description: Create a new IT incident
      operationId: createIncident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentCreate'
      responses:
        '201':
          description: Incident created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /incidents/{incidentId}:
    get:
      tags:
        - Incidents
      summary: Get incident by ID
      description: Retrieve a specific incident by its ID
      operationId: getIncidentById
      parameters:
        - $ref: '#/components/parameters/IncidentIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Incidents
      summary: Update an incident
      description: Update an existing incident (full replacement)
      operationId: updateIncident
      parameters:
        - $ref: '#/components/parameters/IncidentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentUpdate'
      responses:
        '200':
          description: Incident updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Incidents
      summary: Partially update an incident
      description: Update specific fields of an existing incident
      operationId: patchIncident
      parameters:
        - $ref: '#/components/parameters/IncidentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentPatch'
      responses:
        '200':
          description: Incident updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Incidents
      summary: Delete an incident
      description: Delete an incident by its ID
      operationId: deleteIncident
      parameters:
        - $ref: '#/components/parameters/IncidentIdParam'
      responses:
        '204':
          description: Incident deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /incidents/{incidentId}/resolve:
    post:
      tags:
        - Incidents
      summary: Resolve an incident
      description: Mark an incident as resolved with resolution details
      operationId: resolveIncident
      parameters:
        - $ref: '#/components/parameters/IncidentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentResolution'
      responses:
        '200':
          description: Incident resolved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Incident is already resolved or closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /incidents/{incidentId}/close:
    post:
      tags:
        - Incidents
      summary: Close an incident
      description: Close a resolved incident
      operationId: closeIncident
      parameters:
        - $ref: '#/components/parameters/IncidentIdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                closingNotes:
                  type: string
                  description: Optional closing notes
      responses:
        '200':
          description: Incident closed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Incident must be resolved before closing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /incidents/{incidentId}/reopen:
    post:
      tags:
        - Incidents
      summary: Reopen an incident
      description: Reopen a closed or resolved incident
      operationId: reopenIncident
      parameters:
        - $ref: '#/components/parameters/IncidentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  description: Reason for reopening the incident
      responses:
        '200':
          description: Incident reopened successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Incident is not in a reopenable state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /incidents/{incidentId}/assign:
    post:
      tags:
        - Incidents
      summary: Assign users to an incident
      description: Assign one or more users to work on an incident
      operationId: assignIncident
      parameters:
        - $ref: '#/components/parameters/IncidentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - assigneeIds
              properties:
                assigneeIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: List of owner IDs to assign
      responses:
        '200':
          description: Assignees updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /incidents/{incidentId}/comments:
    get:
      tags:
        - Incidents
      summary: Get incident comments
      description: Retrieve all comments for an incident
      operationId: getIncidentComments
      parameters:
        - $ref: '#/components/parameters/IncidentIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Incidents
      summary: Add a comment to an incident
      description: Add a new comment to an incident
      operationId: addIncidentComment
      parameters:
        - $ref: '#/components/parameters/IncidentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '201':
          description: Comment added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /incidents/{incidentId}/timeline:
    get:
      tags:
        - Incidents
      summary: Get incident timeline
      description: Retrieve the audit timeline/history for an incident
      operationId: getIncidentTimeline
      parameters:
        - $ref: '#/components/parameters/IncidentIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimelineEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== OWNERS ====================
...
components:
...
  schemas:
...
    IncidentStatus:
      type: string
      enum:
        - open
        - in_progress
        - on_hold
        - resolved
        - closed
      description: Current status of the incident
    Priority:
      type: string
      enum:
        - low
        - medium
        - high
        - critical
      description: Priority level of the incident
    Severity:
      type: string
      enum:
        - low
        - medium
        - high
        - critical
      description: Severity level indicating business impact
    IncidentCategory:
      type: string
      enum:
        - network
        - hardware
        - software
        - security
        - database
        - cloud_infrastructure
        - application
        - performance
        - access_permissions
        - other
      description: Category of the incident
...
    Incident:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the incident
          readOnly: true
        incidentNumber:
          type: string
          description: Human-readable incident number (e.g., INC-2026-0001)
          readOnly: true
        title:
          type: string
          description: Short title/summary of the incident
          maxLength: 255
        description:
          type: string
          description: Detailed description of the incident
        status:
          $ref: '#/components/schemas/IncidentStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        severity:
          $ref: '#/components/schemas/Severity'
        category:
          $ref: '#/components/schemas/IncidentCategory'
        tags:
          type: array
          items:
            type: string
          description: Tags/labels for categorization
        affectedSystems:
          type: array
          items:
            type: string
          description: List of affected systems/services
        affectedUsers:
          type: integer
          description: Estimated number of affected users
        owner:
          $ref: '#/components/schemas/OwnerSummary'
        assignees:
          type: array
          items:
            $ref: '#/components/schemas/OwnerSummary'
          description: Users assigned to work on the incident
        rootCause:
          type: string
          description: Root cause analysis
          nullable: true
        resolution:
          type: string
          description: Resolution description
          nullable: true
        workaround:
          type: string
          description: Temporary workaround if available
          nullable: true
        githubRepo:
          $ref: '#/components/schemas/GitHubRepo'
          nullable: true
        dueDate:
          type: string
          format: date-time
          description: Expected resolution due date
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the incident was created
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the incident was last updated
          readOnly: true
        acknowledgedAt:
          type: string
          format: date-time
          description: Timestamp when the incident was acknowledged
          nullable: true
          readOnly: true
        resolvedAt:
          type: string
          format: date-time
          description: Timestamp when the incident was resolved
          nullable: true
          readOnly: true
        closedAt:
          type: string
          format: date-time
          description: Timestamp when the incident was closed
          nullable: true
          readOnly: true
        slaBreached:
          type: boolean
          description: Whether the SLA has been breached
          readOnly: true
        timeToAcknowledge:
          type: integer
          description: Time to acknowledge in minutes
          nullable: true
          readOnly: true
        timeToResolve:
          type: integer
          description: Time to resolve in minutes
          nullable: true
          readOnly: true
      required:
        - id
        - incidentNumber
        - title
        - status
        - priority
        - severity
        - category
        - createdAt
        - updatedAt
    IncidentCreate:
      type: object
      required:
        - title
        - description
        - priority
        - severity
        - category
        - ownerId
      properties:
        title:
          type: string
          description: Short title/summary of the incident
          maxLength: 255
        description:
          type: string
          description: Detailed description of the incident
        priority:
          $ref: '#/components/schemas/Priority'
        severity:
          $ref: '#/components/schemas/Severity'
        category:
          $ref: '#/components/schemas/IncidentCategory'
        tags:
          type: array
          items:
            type: string
          description: Tags/labels for categorization
        affectedSystems:
          type: array
          items:
            type: string
          description: List of affected systems/services
        affectedUsers:
          type: integer
          description: Estimated number of affected users
        ownerId:
          type: string
          format: uuid
          description: ID of the owner responsible for this incident
        assigneeIds:
          type: array
          items:
            type: string
            format: uuid
          description: IDs of users to assign to this incident
        workaround:
          type: string
          description: Temporary workaround if available
        githubRepo:
          $ref: '#/components/schemas/GitHubRepoInput'
        dueDate:
          type: string
          format: date-time
          description: Expected resolution due date
    IncidentUpdate:
      type: object
      required:
        - title
        - description
        - priority
        - severity
        - category
        - ownerId
      properties:
        title:
          type: string
          description: Short title/summary of the incident
          maxLength: 255
        description:
          type: string
          description: Detailed description of the incident
        status:
          $ref: '#/components/schemas/IncidentStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        severity:
          $ref: '#/components/schemas/Severity'
        category:
          $ref: '#/components/schemas/IncidentCategory'
        tags:
          type: array
          items:
            type: string
        affectedSystems:
          type: array
          items:
            type: string
        affectedUsers:
          type: integer
        ownerId:
          type: string
          format: uuid
        assigneeIds:
          type: array
          items:
            type: string
            format: uuid
        rootCause:
          type: string
        resolution:
          type: string
        workaround:
          type: string
        githubRepo:
          $ref: '#/components/schemas/GitHubRepoInput'
        dueDate:
          type: string
          format: date-time
    IncidentPatch:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        status:
          $ref: '#/components/schemas/IncidentStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        severity:
          $ref: '#/components/schemas/Severity'
        category:
          $ref: '#/components/schemas/IncidentCategory'
        tags:
          type: array
          items:
            type: string
        affectedSystems:
          type: array
          items:
            type: string
        affectedUsers:
          type: integer
        ownerId:
          type: string
          format: uuid
        assigneeIds:
          type: array
          items:
            type: string
            format: uuid
        rootCause:
          type: string
        resolution:
          type: string
        workaround:
          type: string
        githubRepo:
          $ref: '#/components/schemas/GitHubRepoInput'
        dueDate:
          type: string
          format: date-time
    IncidentResolution:
      type: object
      required:
        - resolution
      properties:
        rootCause:
          type: string
          description: Root cause analysis
        resolution:
          type: string
          description: Description of how the incident was resolved
    IncidentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Incident'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ----- Owner -----
...
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: Detailed validation errors
        traceId:
          type: string
          description: Trace ID for debugging
        timestamp:
          type: string
          format: date-time

  # ==================== PARAMETERS ====================
  parameters:
    IncidentIdParam:
      name: incidentId
      in: path
      required: true
      description: Unique identifier of the incident
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sortBy
      in: query
      description: Field to sort by
      schema:
        type: string
        enum:
          - createdAt
          - updatedAt
          - priority
          - severity
          - status
          - dueDate
          - title
        default: createdAt
    SortOrderParam:
      name: sortOrder
      in: query
      description: Sort order
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc

  # ==================== RESPONSES ====================
  responses:
    BadRequest:
      description: Bad request - invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: Unprocessable entity - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # ==================== SECURITY ====================
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
security:
  - bearerAuth: []
  - apiKeyAuth: []